Move test_recommend_n_examples back to test_full
